Add files via upload
=> 𝙄𝙩 𝙞𝙣𝙘𝙡𝙪𝙙𝙚𝙨 importing pandas, matplotlib, seaborn,numpy, etc.

- .csv file reading

- data visualisation concepts

- histogram representation of data,etc.



Linear regression is a statistical method that is used to predict a continuous dependent variable(target variable) based on one or more independent variables(predictor variables). 

This technique assumes a linear relationship between the dependent and independent variables, which implies that the dependent variable changes proportionally with changes in the independent variables. 

In other words, linear regression is used to determine the extent to which one or more variables can predict the value of the dependent variable.